pyproject.toml
[project]
name = "173-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.41.0",
    "gunicorn>=21.0.0",
]
